Fix C4267 size_t narrowing warnings (round 2)
- xc_task.hpp: cast points.size() to int32_t in both merge_with overloads
- basisset.hpp: cast size() to int32_t in nshells() return
- misc.hpp: cast std::min(A.size(), B.size()) to uint32_t
- integrate_den.hpp: cast task sizes to int32_t
- reference_local_host_work_driver.cxx: add inbf locals and cast npts/nbe_cart
  at sph_trans/submat/inc_by_submat_atomic call boundaries; fix ioff = i*nbf
  (size_t) -> i*inbf (int32_t*int32_t)

+++ include/gauxc/xc_task.hpp
@@ -66,7 +66,7 @@ struct XCTask {
       GAUXC_GENERIC_EXCEPTION("Cannot Perform Requested Merge: Incompatible Tasks");
     points.insert( points.end(), other.points.begin(), other.points.end() );
     weights.insert( weights.end(), other.weights.begin(), other.weights.end() );
-    npts = points.size();
+    npts = static_cast<int32_t>(points.size());
   }
 
   template <typename TaskIt>
@@ -91,9 +91,8 @@ struct XCTask {
       weights_it = std::copy( it->weights.begin(), it->weights.end(), weights_it );
     }
 
-    npts = points.size();
+    npts = static_cast<int32_t>(points.size());
   }
-
 
   inline bool equiv_with( const XCTask& other ) const {
     return iParent == other.iParent and
